requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
requests
python-multipart
ollama
alembic
python-dotenv
pandas
openpyxl
